Run updater before starting the controllers.
